Updates to code and some visual GUI changes for coloring.
Now takes a snapshot of some of the settings at beginning of run and
uses that snapshot (and saves it). This is good for starting at
intermediate processing steps. Also made a couple of relevant objects
serializable to allow saving of states along throughout processing.
Added ability to toggle false coloring on and off. Now detects updates.